Fix lifecycle aggregate state inputs
